requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
supabase==2.0.0
flask-socketio
simple-websocket
